requirements.txt
# Core dependencies
openai>=1.0.0
langchain>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# API integrations
requests>=2.28.0
aiohttp>=3.8.0
websockets>=10.0

# Financial analysis
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
yfinance>=0.2.0
alpha_vantage>=2.3.1

# GitHub integration
PyGithub>=1.58.0

# User interfaces
streamlit>=1.22.0
python-telegram-bot>=20.0

# Database and memory
sqlalchemy>=2.0.0
redis>=4.5.0

# Utilities
tqdm>=4.65.0
colorama>=0.4.6
loguru>=0.7.0

# Deployment
gunicorn>=21.0.0
uvicorn>=0.22.0

# Web framework
fastapi>=0.95.0

# Testing
pytest>=7.3.0

# Documentation
sphinx>=7.0.0
